Squashed kgsl updates
msm_kgsl: Move kgsl_sharedmem_init to module load

kgsl_sharedmem_init/kgsl_sharedmem_close only need to be called once,
so move them to the the module load and close functions so they can be
available through the life of the module.

msm_kgsl: Optimize TLB flushing

Only flush TLB when it is absolutely nessasary as opposed to
flushing it on every map and unmap

msm: kgsl: Fix a potential subtraction overflow issue

CRs-fixed: 262785
Change-Id: Ic0de06f59fe4b37056d89e3672a8320ec814a418
Signed-off-by: Wei Zou <wzou@codeaurora.org>

msm_kgsl: Reuse vmalloced memory buffers

Memory buffers that are vmalloced and can be freed are put
into a free list and reused later if an allocation of same size
is required.

Change-Id: Ia7387ee8235cc192d20646cfef2510a9820d464b
Signed-off-by: Shubhraprakash Das <sadas@quicinc.com>